decouple PharFileFinder
